server/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.11
django-simpleui==2024.4.1
exceptiongroup==1.2.1
fastapi==0.68.1
google-api-python-client==2.86.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
idna==3.7
iniconfig==2.0.0
lxml==5.2.1
packaging==24.0
pillow==10.3.0
pluggy==1.5.0
pytest==8.2.0
python-dotenv==0.19.0
requests==2.31.0
soupsieve==2.5
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.15.0
